7228: Reorganized files
