+++ assets/client/components/screen.jsx
@@ -2,9 +2,9 @@ import { useEffect, useRef } from "react";
 import SunCalc from "suncalc";
 import { createGrid } from "../../shared/grid-generator/grid-generator";
 import Region from "./region.jsx";
-import logger from "../logger/logger";
+import logger from "../logger";
 import TouchRegion from "./touch-region.jsx";
-import ClientConfigLoader from "../util/client-config-loader.js";
+import ClientConfigLoader from "../core/client-config-loader.js";
 import constants from "../util/constants";
 import "./screen.scss";
