+++ assets/client/core/app-storage.js
