+++ assets/client/logger.js
