+++ assets/client/service/release-service.js
@@ -1,8 +1,8 @@
-import ClientConfigLoader from "../util/client-config-loader.js";
+import ClientConfigLoader from "../core/client-config-loader.js";
 import defaults from "../util/defaults";
 import idFromPath from "../util/id-from-path";
-import appStorage from "../util/app-storage";
-import logger from "../logger/logger";
+import appStorage from "../core/app-storage";
+import logger from "../logger";
 import statusService from "./status-service";
 import constants from "../util/constants";
 import releaseLoader from "../../shared/release-loader.js";
